Remove user-defined slash commands feature
