+++ crates/agent_ui/src/acp/message_editor.rs
@@ -9,7 +9,6 @@ use crate::{
     mention_set::{
         Mention, MentionImage, MentionSet, insert_crease_for_mention, paste_images_as_context,
     },
-    user_slash_command::{self, CommandLoadError, UserSlashCommand},
 };
 use acp_thread::{AgentSessionInfo, MentionUri};
 use agent::ThreadStore;
@@ -22,7 +21,6 @@ use editor::{
     MultiBufferSnapshot, ToOffset, actions::Paste, code_context_menus::CodeContextMenu,
     scroll::Autoscroll,
 };
-use feature_flags::{FeatureFlagAppExt as _, UserSlashCommandsFeatureFlag};
 use futures::{FutureExt as _, future::join_all};
 use gpui::{
     AppContext, ClipboardEntry, Context, Entity, EventEmitter, FocusHandle, Focusable, ImageFormat,
@@ -40,25 +38,12 @@ use util::{ResultExt, debug_panic};
 use workspace::{CollaboratorId, Workspace};
 use zed_actions::agent::{Chat, PasteRaw};
 
-enum UserSlashCommands {
-    Cached {
-        commands: collections::HashMap<String, user_slash_command::UserSlashCommand>,
-        errors: Vec<user_slash_command::CommandLoadError>,
-    },
-    FromFs {
-        fs: Arc<dyn fs::Fs>,
-        worktree_roots: Vec<std::path::PathBuf>,
-    },
-}
-
 pub struct MessageEditor {
     mention_set: Entity<MentionSet>,
     editor: Entity<Editor>,
     workspace: WeakEntity<Workspace>,
     prompt_capabilities: Rc<RefCell<acp::PromptCapabilities>>,
     available_commands: Rc<RefCell<Vec<acp::AvailableCommand>>>,
-    cached_user_commands: Rc<RefCell<collections::HashMap<String, UserSlashCommand>>>,
-    cached_user_command_errors: Rc<RefCell<Vec<CommandLoadError>>>,
     agent_name: SharedString,
     thread_store: Option<Entity<ThreadStore>>,
     _subscriptions: Vec<Subscription>,
@@ -107,34 +92,12 @@ impl PromptCompletionProviderDelegate for Entity<MessageEditor> {
                 name: cmd.name.clone().into(),
                 description: cmd.description.clone().into(),
                 requires_argument: cmd.input.is_some(),
-                source: crate::completion_provider::CommandSource::Server,
             })
             .collect()
     }
 
     fn confirm_command(&self, cx: &mut App) {
         self.update(cx, |this, cx| this.send(cx));
-    }
-
-    fn cached_user_commands(
-        &self,
-        cx: &App,
-    ) -> Option<collections::HashMap<String, UserSlashCommand>> {
-        let commands = self.read(cx).cached_user_commands.borrow();
-        if commands.is_empty() {
-            None
-        } else {
-            Some(commands.clone())
-        }
-    }
-
-    fn cached_user_command_errors(&self, cx: &App) -> Option<Vec<CommandLoadError>> {
-        let errors = self.read(cx).cached_user_command_errors.borrow();
-        if errors.is_empty() {
-            None
-        } else {
-            Some(errors.clone())
-        }
     }
 }
 
@@ -147,42 +110,6 @@ impl MessageEditor {
         prompt_store: Option<Entity<PromptStore>>,
         prompt_capabilities: Rc<RefCell<acp::PromptCapabilities>>,
         available_commands: Rc<RefCell<Vec<acp::AvailableCommand>>>,
-        agent_name: SharedString,
-        placeholder: &str,
-        mode: EditorMode,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> Self {
-        let cached_user_commands = Rc::new(RefCell::new(collections::HashMap::default()));
-        let cached_user_command_errors = Rc::new(RefCell::new(Vec::new()));
-        Self::new_with_cache(
-            workspace,
-            project,
-            thread_store,
-            history,
-            prompt_store,
-            prompt_capabilities,
-            available_commands,
-            cached_user_commands,
-            cached_user_command_errors,
-            agent_name,
-            placeholder,
-            mode,
-            window,
-            cx,
-        )
-    }
-
-    pub fn new_with_cache(
-        workspace: WeakEntity<Workspace>,
-        project: WeakEntity<Project>,
-        thread_store: Option<Entity<ThreadStore>>,
-        history: WeakEntity<AcpThreadHistory>,
-        prompt_store: Option<Entity<PromptStore>>,
-        prompt_capabilities: Rc<RefCell<acp::PromptCapabilities>>,
-        available_commands: Rc<RefCell<Vec<acp::AvailableCommand>>>,
-        cached_user_commands: Rc<RefCell<collections::HashMap<String, UserSlashCommand>>>,
-        cached_user_command_errors: Rc<RefCell<Vec<CommandLoadError>>>,
         agent_name: SharedString,
         placeholder: &str,
         mode: EditorMode,
@@ -293,8 +220,6 @@ impl MessageEditor {
             workspace,
             prompt_capabilities,
             available_commands,
-            cached_user_commands,
-            cached_user_command_errors,
             agent_name,
             thread_store,
             _subscriptions: subscriptions,
@@ -306,14 +231,10 @@ impl MessageEditor {
         &mut self,
         prompt_capabilities: Rc<RefCell<acp::PromptCapabilities>>,
         available_commands: Rc<RefCell<Vec<acp::AvailableCommand>>>,
-        cached_user_commands: Rc<RefCell<collections::HashMap<String, UserSlashCommand>>>,
-        cached_user_command_errors: Rc<RefCell<Vec<CommandLoadError>>>,
         _cx: &mut Context<Self>,
     ) {
         self.prompt_capabilities = prompt_capabilities;
         self.available_commands = available_commands;
-        self.cached_user_commands = cached_user_commands;
-        self.cached_user_command_errors = cached_user_command_errors;
     }
 
     fn command_hint(&self, snapshot: &MultiBufferSnapshot) -> Option<Inlay> {
@@ -478,46 +399,9 @@ impl MessageEditor {
         full_mention_content: bool,
         cx: &mut Context<Self>,
     ) -> Task<Result<(Vec<acp::ContentBlock>, Vec<Entity<Buffer>>)>> {
-        self.contents_with_cache(full_mention_content, None, None, cx)
-    }
-
-    pub fn contents_with_cache(
-        &self,
-        full_mention_content: bool,
-        cached_user_commands: Option<
-            collections::HashMap<String, user_slash_command::UserSlashCommand>,
-        >,
-        cached_user_command_errors: Option<Vec<user_slash_command::CommandLoadError>>,
-        cx: &mut Context<Self>,
-    ) -> Task<Result<(Vec<acp::ContentBlock>, Vec<Entity<Buffer>>)>> {
         let text = self.editor.read(cx).text(cx);
         let available_commands = self.available_commands.borrow().clone();
         let agent_name = self.agent_name.clone();
-
-        let user_slash_commands = if !cx.has_flag::<UserSlashCommandsFeatureFlag>() {
-            UserSlashCommands::Cached {
-                commands: collections::HashMap::default(),
-                errors: Vec::new(),
-            }
-        } else if let Some(cached) = cached_user_commands {
-            UserSlashCommands::Cached {
-                commands: cached,
-                errors: cached_user_command_errors.unwrap_or_default(),
-            }
-        } else if let Some(workspace) = self.workspace.upgrade() {
-            let fs = workspace.read(cx).project().read(cx).fs().clone();
-            let worktree_roots: Vec<std::path::PathBuf> = workspace
-                .read(cx)
-                .visible_worktrees(cx)
-                .map(|worktree| worktree.read(cx).abs_path().to_path_buf())
-                .collect();
-            UserSlashCommands::FromFs { fs, worktree_roots }
-        } else {
-            UserSlashCommands::Cached {
-                commands: collections::HashMap::default(),
-                errors: Vec::new(),
-            }
-        };
 
         let contents = self
             .mention_set
@@ -526,58 +410,7 @@ impl MessageEditor {
         let supports_embedded_context = self.prompt_capabilities.borrow().embedded_context;
 
         cx.spawn(async move |_, cx| {
-            let (mut user_commands, mut user_command_errors) = match user_slash_commands {
-                UserSlashCommands::Cached { commands, errors } => (commands, errors),
-                UserSlashCommands::FromFs { fs, worktree_roots } => {
-                    let load_result =
-                        user_slash_command::load_all_commands_async(&fs, &worktree_roots).await;
-
-                    (
-                        user_slash_command::commands_to_map(&load_result.commands),
-                        load_result.errors,
-                    )
-                }
-            };
-
-            let server_command_names = available_commands
-                .iter()
-                .map(|command| command.name.clone())
-                .collect::<HashSet<_>>();
-            user_slash_command::apply_server_command_conflicts_to_map(
-                &mut user_commands,
-                &mut user_command_errors,
-                &server_command_names,
-            );
-
-            // Check if the user is trying to use an errored slash command.
-            // If so, report the error to the user.
-            if let Some(parsed) = user_slash_command::try_parse_user_command(&text) {
-                for error in &user_command_errors {
-                    if let Some(error_cmd_name) = error.command_name() {
-                        if error_cmd_name == parsed.name {
-                            return Err(anyhow::anyhow!(
-                                "Failed to load /{}: {}",
-                                parsed.name,
-                                error.message
-                            ));
-                        }
-                    }
-                }
-            }
-            // Errors for commands that don't match the user's input are silently ignored here,
-            // since the user will see them via the error callout in the thread view.
-
-            // Check if this is a user-defined slash command and expand it
-            match user_slash_command::try_expand_from_commands(&text, &user_commands) {
-                Ok(Some(expanded)) => return Ok((vec![expanded.into()], Vec::new())),
-                Err(err) => return Err(err),
-                Ok(None) => {} // Not a user command, continue with normal processing
-            }
-
-            if let Err(err) = Self::validate_slash_commands(&text, &available_commands, &agent_name)
-            {
-                return Err(err);
-            }
+            Self::validate_slash_commands(&text, &available_commands, &agent_name)?;
 
             let contents = contents.await?;
             let mut all_tracked_buffers = Vec::new();
@@ -1497,14 +1330,12 @@ mod tests {
 
         let message_editor = cx.update(|window, cx| {
             cx.new(|cx| {
-                MessageEditor::new_with_cache(
+                MessageEditor::new(
                     workspace.downgrade(),
                     project.downgrade(),
                     thread_store.clone(),
                     history.downgrade(),
                     None,
-                    Default::default(),
-                    Default::default(),
                     Default::default(),
                     Default::default(),
                     "Test Agent".into(),
@@ -1574,9 +1405,7 @@ mod tests {
         });
 
         let (content, _) = message_editor
-            .update(cx, |message_editor, cx| {
-                message_editor.contents_with_cache(false, None, None, cx)
-            })
+            .update(cx, |message_editor, cx| message_editor.contents(false, cx))
             .await
             .unwrap();
 
@@ -1614,7 +1443,7 @@ mod tests {
         let workspace_handle = workspace.downgrade();
         let message_editor = workspace.update_in(cx, |_, window, cx| {
             cx.new(|cx| {
-                MessageEditor::new_with_cache(
+                MessageEditor::new(
                     workspace_handle.clone(),
                     project.downgrade(),
                     thread_store.clone(),
@@ -1622,8 +1451,6 @@ mod tests {
                     None,
                     prompt_capabilities.clone(),
                     available_commands.clone(),
-                    Default::default(),
-                    Default::default(),
                     "Claude Code".into(),
                     "Test",
                     EditorMode::AutoHeight {
@@ -1643,9 +1470,7 @@ mod tests {
         });
 
         let contents_result = message_editor
-            .update(cx, |message_editor, cx| {
-                message_editor.contents_with_cache(false, None, None, cx)
-            })
+            .update(cx, |message_editor, cx| message_editor.contents(false, cx))
             .await;
 
         // Should fail because available_commands is empty (no commands supported)
@@ -1663,9 +1488,7 @@ mod tests {
         });
 
         let contents_result = message_editor
-            .update(cx, |message_editor, cx| {
-                message_editor.contents_with_cache(false, None, None, cx)
-            })
+            .update(cx, |message_editor, cx| message_editor.contents(false, cx))
             .await;
 
         assert!(contents_result.is_err());
@@ -1680,9 +1503,7 @@ mod tests {
         });
 
         let contents_result = message_editor
-            .update(cx, |message_editor, cx| {
-                message_editor.contents_with_cache(false, None, None, cx)
-            })
+            .update(cx, |message_editor, cx| message_editor.contents(false, cx))
             .await;
 
         // Should succeed because /help is in available_commands
@@ -1694,9 +1515,7 @@ mod tests {
         });
 
         let (content, _) = message_editor
-            .update(cx, |message_editor, cx| {
-                message_editor.contents_with_cache(false, None, None, cx)
-            })
+            .update(cx, |message_editor, cx| message_editor.contents(false, cx))
             .await
             .unwrap();
 
@@ -1714,9 +1533,7 @@ mod tests {
 
         // The @ mention functionality should not be affected
         let (content, _) = message_editor
-            .update(cx, |message_editor, cx| {
-                message_editor.contents_with_cache(false, None, None, cx)
-            })
+            .update(cx, |message_editor, cx| message_editor.contents(false, cx))
             .await
             .unwrap();
 
@@ -1789,7 +1606,7 @@ mod tests {
         let editor = workspace.update_in(&mut cx, |workspace, window, cx| {
             let workspace_handle = cx.weak_entity();
             let message_editor = cx.new(|cx| {
-                MessageEditor::new_with_cache(
+                MessageEditor::new(
                     workspace_handle,
                     project.downgrade(),
                     thread_store.clone(),
@@ -1797,8 +1614,6 @@ mod tests {
                     None,
                     prompt_capabilities.clone(),
                     available_commands.clone(),
-                    Default::default(),
-                    Default::default(),
                     "Test Agent".into(),
                     "Test",
                     EditorMode::AutoHeight {
@@ -2015,15 +1830,13 @@ mod tests {
         let (message_editor, editor) = workspace.update_in(&mut cx, |workspace, window, cx| {
             let workspace_handle = cx.weak_entity();
             let message_editor = cx.new(|cx| {
-                MessageEditor::new_with_cache(
+                MessageEditor::new(
                     workspace_handle,
                     project.downgrade(),
                     Some(thread_store),
                     history.downgrade(),
                     None,
                     prompt_capabilities.clone(),
-                    Default::default(),
-                    Default::default(),
                     Default::default(),
                     "Test Agent".into(),
                     "Test",
@@ -2510,14 +2323,12 @@ mod tests {
 
         let message_editor = cx.update(|window, cx| {
             cx.new(|cx| {
-                let editor = MessageEditor::new_with_cache(
+                let editor = MessageEditor::new(
                     workspace.downgrade(),
                     project.downgrade(),
                     thread_store.clone(),
                     history.downgrade(),
                     None,
-                    Default::default(),
-                    Default::default(),
                     Default::default(),
                     Default::default(),
                     "Test Agent".into(),
@@ -2621,14 +2432,12 @@ mod tests {
 
         let message_editor = cx.update(|window, cx| {
             cx.new(|cx| {
-                let mut editor = MessageEditor::new_with_cache(
+                let mut editor = MessageEditor::new(
                     workspace.downgrade(),
                     project.downgrade(),
                     thread_store.clone(),
                     history.downgrade(),
                     None,
-                    Default::default(),
-                    Default::default(),
                     Default::default(),
                     Default::default(),
                     "Test Agent".into(),
@@ -2703,14 +2512,12 @@ mod tests {
 
         let message_editor = cx.update(|window, cx| {
             cx.new(|cx| {
-                let mut editor = MessageEditor::new_with_cache(
+                let mut editor = MessageEditor::new(
                     workspace.downgrade(),
                     project.downgrade(),
                     thread_store.clone(),
                     history.downgrade(),
                     None,
-                    Default::default(),
-                    Default::default(),
                     Default::default(),
                     Default::default(),
                     "Test Agent".into(),
@@ -2756,14 +2563,12 @@ mod tests {
 
         let message_editor = cx.update(|window, cx| {
             cx.new(|cx| {
-                MessageEditor::new_with_cache(
+                MessageEditor::new(
                     workspace.downgrade(),
                     project.downgrade(),
                     thread_store.clone(),
                     history.downgrade(),
                     None,
-                    Default::default(),
-                    Default::default(),
                     Default::default(),
                     Default::default(),
                     "Test Agent".into(),
@@ -2812,14 +2617,12 @@ mod tests {
 
         let message_editor = cx.update(|window, cx| {
             cx.new(|cx| {
-                MessageEditor::new_with_cache(
+                MessageEditor::new(
                     workspace.downgrade(),
                     project.downgrade(),
                     thread_store.clone(),
                     history.downgrade(),
                     None,
-                    Default::default(),
-                    Default::default(),
                     Default::default(),
                     Default::default(),
                     "Test Agent".into(),
@@ -2869,14 +2672,12 @@ mod tests {
 
         let message_editor = cx.update(|window, cx| {
             cx.new(|cx| {
-                MessageEditor::new_with_cache(
+                MessageEditor::new(
                     workspace.downgrade(),
                     project.downgrade(),
                     thread_store.clone(),
                     history.downgrade(),
                     None,
-                    Default::default(),
-                    Default::default(),
                     Default::default(),
                     Default::default(),
                     "Test Agent".into(),
@@ -2899,9 +2700,7 @@ mod tests {
         });
 
         let (content, _) = message_editor
-            .update(cx, |message_editor, cx| {
-                message_editor.contents_with_cache(false, None, None, cx)
-            })
+            .update(cx, |message_editor, cx| message_editor.contents(false, cx))
             .await
             .unwrap();
 
@@ -2938,14 +2737,12 @@ mod tests {
         let (message_editor, editor) = workspace.update_in(cx, |workspace, window, cx| {
             let workspace_handle = cx.weak_entity();
             let message_editor = cx.new(|cx| {
-                MessageEditor::new_with_cache(
+                MessageEditor::new(
                     workspace_handle,
                     project.downgrade(),
                     thread_store.clone(),
                     history.downgrade(),
                     None,
-                    Default::default(),
-                    Default::default(),
                     Default::default(),
                     Default::default(),
                     "Test Agent".into(),
@@ -2982,9 +2779,7 @@ mod tests {
         });
 
         let content = message_editor
-            .update(cx, |editor, cx| {
-                editor.contents_with_cache(false, None, None, cx)
-            })
+            .update(cx, |editor, cx| editor.contents(false, cx))
             .await
             .unwrap()
             .0;
@@ -3011,9 +2806,7 @@ mod tests {
         });
 
         let content = message_editor
-            .update(cx, |editor, cx| {
-                editor.contents_with_cache(false, None, None, cx)
-            })
+            .update(cx, |editor, cx| editor.contents(false, cx))
             .await
             .unwrap()
             .0;
@@ -3104,14 +2897,12 @@ mod tests {
         let message_editor = workspace.update_in(&mut cx, |workspace, window, cx| {
             let workspace_handle = cx.weak_entity();
             let message_editor = cx.new(|cx| {
-                MessageEditor::new_with_cache(
+                MessageEditor::new(
                     workspace_handle,
                     project.downgrade(),
                     thread_store.clone(),
                     history.downgrade(),
                     None,
-                    Default::default(),
-                    Default::default(),
                     Default::default(),
                     Default::default(),
                     "Test Agent".into(),
@@ -3221,14 +3012,12 @@ mod tests {
         let (message_editor, editor) = workspace.update_in(&mut cx, |workspace, window, cx| {
             let workspace_handle = cx.weak_entity();
             let message_editor = cx.new(|cx| {
-                MessageEditor::new_with_cache(
+                MessageEditor::new(
                     workspace_handle,
                     project.downgrade(),
                     Some(thread_store),
                     history.downgrade(),
                     None,
-                    Default::default(),
-                    Default::default(),
                     Default::default(),
                     Default::default(),
                     "Test Agent".into(),
