+++ crates/agent_ui/src/completion_provider.rs
@@ -5,14 +5,11 @@ use std::sync::Arc;
 use std::sync::atomic::AtomicBool;
 
 use crate::acp::AcpThreadHistory;
-use crate::user_slash_command::{self, CommandLoadError, UserSlashCommand};
 use acp_thread::{AgentSessionInfo, MentionUri};
 use anyhow::Result;
-use collections::{HashMap, HashSet};
 use editor::{
     CompletionProvider, Editor, ExcerptId, code_context_menus::COMPLETION_MENU_MAX_WIDTH,
 };
-use feature_flags::{FeatureFlagAppExt as _, UserSlashCommandsFeatureFlag};
 use futures::FutureExt as _;
 use fuzzy::{PathMatch, StringMatch, StringMatchCandidate};
 use gpui::{App, BackgroundExecutor, Entity, SharedString, Task, WeakEntity};
@@ -192,18 +189,6 @@ pub struct AvailableCommand {
     pub name: Arc<str>,
     pub description: Arc<str>,
     pub requires_argument: bool,
-    pub source: CommandSource,
-}
-
-/// The source of a slash command, used to differentiate UI behavior.
-#[derive(Debug, Clone, PartialEq)]
-pub enum CommandSource {
-    /// Command provided by the ACP server
-    Server,
-    /// User-defined command from a markdown file
-    UserDefined { template: Arc<str> },
-    /// User-defined command that failed to load
-    UserDefinedError { error_message: Arc<str> },
 }
 
 pub trait PromptCompletionProviderDelegate: Send + Sync + 'static {
@@ -215,18 +200,6 @@ pub trait PromptCompletionProviderDelegate: Send + Sync + 'static {
 
     fn available_commands(&self, cx: &App) -> Vec<AvailableCommand>;
     fn confirm_command(&self, cx: &mut App);
-
-    /// Returns cached user-defined slash commands, if available.
-    /// Default implementation returns None, meaning commands will be loaded from disk.
-    fn cached_user_commands(&self, _cx: &App) -> Option<HashMap<String, UserSlashCommand>> {
-        None
-    }
-
-    /// Returns cached errors from loading user-defined slash commands, if available.
-    /// Default implementation returns None.
-    fn cached_user_command_errors(&self, _cx: &App) -> Option<Vec<CommandLoadError>> {
-        None
-    }
 }
 
 pub struct PromptCompletionProvider<T: PromptCompletionProviderDelegate> {
@@ -803,111 +776,11 @@ impl<T: PromptCompletionProviderDelegate> PromptCompletionProvider<T> {
 
     fn search_slash_commands(&self, query: String, cx: &mut App) -> Task<Vec<AvailableCommand>> {
         let commands = self.source.available_commands(cx);
-        let server_command_names = commands
-            .iter()
-            .map(|command| command.name.as_ref().to_string())
-            .collect::<HashSet<_>>();
-
-        // Try to use cached user commands and errors first
-        let cached_user_commands = if cx.has_flag::<UserSlashCommandsFeatureFlag>() {
-            self.source.cached_user_commands(cx)
-        } else {
-            None
-        };
-
-        let cached_user_command_errors = if cx.has_flag::<UserSlashCommandsFeatureFlag>() {
-            self.source.cached_user_command_errors(cx)
-        } else {
-            None
-        };
-
-        // Get fs and worktree roots for async command loading (only if not cached)
-        let (fs, worktree_roots) =
-            if cached_user_commands.is_none() && cx.has_flag::<UserSlashCommandsFeatureFlag>() {
-                let workspace = self.workspace.upgrade();
-                let fs = workspace
-                    .as_ref()
-                    .map(|w| w.read(cx).project().read(cx).fs().clone());
-                let roots: Vec<std::path::PathBuf> = workspace
-                    .map(|workspace| {
-                        workspace
-                            .read(cx)
-                            .visible_worktrees(cx)
-                            .map(|worktree| worktree.read(cx).abs_path().to_path_buf())
-                            .collect()
-                    })
-                    .unwrap_or_default();
-                (fs, roots)
-            } else {
-                (None, Vec::new())
-            };
+        if commands.is_empty() {
+            return Task::ready(Vec::new());
+        }
 
         cx.spawn(async move |cx| {
-            let mut commands = commands;
-
-            // Use cached commands/errors if available, otherwise load from disk
-            let (mut user_commands, mut user_command_errors): (
-                Vec<UserSlashCommand>,
-                Vec<CommandLoadError>,
-            ) = if let Some(cached) = cached_user_commands {
-                let errors = cached_user_command_errors.unwrap_or_default();
-                (cached.into_values().collect(), errors)
-            } else if let Some(fs) = fs {
-                let load_result =
-                    crate::user_slash_command::load_all_commands_async(&fs, &worktree_roots).await;
-
-                (load_result.commands, load_result.errors)
-            } else {
-                (Vec::new(), Vec::new())
-            };
-
-            user_slash_command::apply_server_command_conflicts(
-                &mut user_commands,
-                &mut user_command_errors,
-                &server_command_names,
-            );
-
-            let conflicting_names: HashSet<String> = user_command_errors
-                .iter()
-                .filter_map(|error| error.command_name())
-                .filter(|name| server_command_names.contains(name))
-                .collect();
-
-            if !conflicting_names.is_empty() {
-                commands.retain(|command| !conflicting_names.contains(command.name.as_ref()));
-            }
-
-            for cmd in user_commands {
-                commands.push(AvailableCommand {
-                    name: cmd.name.clone(),
-                    description: cmd.description().into(),
-                    requires_argument: cmd.requires_arguments(),
-                    source: CommandSource::UserDefined {
-                        template: cmd.template.clone(),
-                    },
-                });
-            }
-
-            // Add errored commands so they show up in autocomplete with error indication.
-            // Errors for commands that don't match the query will be silently ignored here
-            // since the user will see them via the error callout in the thread view.
-            for error in user_command_errors {
-                if let Some(name) = error.command_name() {
-                    commands.push(AvailableCommand {
-                        name: name.into(),
-                        description: "".into(),
-                        requires_argument: false,
-                        source: CommandSource::UserDefinedError {
-                            error_message: error.message.into(),
-                        },
-                    });
-                }
-            }
-
-            if commands.is_empty() {
-                return Vec::new();
-            }
-
             let candidates = commands
                 .iter()
                 .enumerate()
@@ -1264,20 +1137,7 @@ impl<T: PromptCompletionProviderDelegate> CompletionProvider for PromptCompletio
                         .await
                         .into_iter()
                         .map(|command| {
-                            let is_error =
-                                matches!(command.source, CommandSource::UserDefinedError { .. });
-
-                            // For errored commands, show the name with "(load error)" suffix
-                            let label_text = if is_error {
-                                format!("{} (load error)", command.name)
-                            } else {
-                                command.name.to_string()
-                            };
-
-                            // For errored commands, we don't want to insert anything useful
-                            let new_text = if is_error {
-                                format!("/{}", command.name)
-                            } else if let Some(argument) = argument.as_ref() {
+                            let new_text = if let Some(argument) = argument.as_ref() {
                                 format!("/{} {}", command.name, argument)
                             } else {
                                 format!("/{} ", command.name)
@@ -1286,72 +1146,21 @@ impl<T: PromptCompletionProviderDelegate> CompletionProvider for PromptCompletio
                             let is_missing_argument =
                                 command.requires_argument && argument.is_none();
 
-                            // For errored commands, use a deprecated-style label to indicate the error
-                            let label = if is_error {
-                                // Create a label where the command name portion has a highlight
-                                // that will be rendered with strikethrough by the completion menu
-                                // (similar to deprecated LSP completions)
-                                CodeLabel::plain(label_text, None)
-                            } else {
-                                CodeLabel::plain(label_text, None)
-                            };
-
-                            // For errored commands, show the error message in documentation
-                            let documentation =
-                                if let CommandSource::UserDefinedError { error_message } =
-                                    &command.source
-                                {
-                                    Some(CompletionDocumentation::MultiLinePlainText(
-                                        error_message.to_string().into(),
-                                    ))
-                                } else if !command.description.is_empty() {
-                                    Some(CompletionDocumentation::MultiLinePlainText(
-                                        command.description.to_string().into(),
-                                    ))
-                                } else {
-                                    None
-                                };
-
-                            // For errored commands, use a red X icon
-                            let icon_path = if is_error {
-                                Some(IconName::XCircle.path().into())
-                            } else {
-                                None
-                            };
-
                             Completion {
                                 replace_range: source_range.clone(),
                                 new_text,
-                                label,
-                                documentation,
-                                source: if is_error {
-                                    // Use a custom source that marks this as deprecated/errored
-                                    // so the completion menu renders it with strikethrough
-                                    project::CompletionSource::Lsp {
-                                        insert_range: None,
-                                        server_id: language::LanguageServerId(0),
-                                        lsp_completion: Box::new(lsp::CompletionItem {
-                                            label: command.name.to_string(),
-                                            deprecated: Some(true),
-                                            ..Default::default()
-                                        }),
-                                        lsp_defaults: None,
-                                        resolved: true,
-                                    }
-                                } else {
-                                    project::CompletionSource::Custom
-                                },
-                                icon_path,
+                                label: CodeLabel::plain(command.name.to_string(), None),
+                                documentation: Some(CompletionDocumentation::MultiLinePlainText(
+                                    command.description.into(),
+                                )),
+                                source: project::CompletionSource::Custom,
+                                icon_path: None,
                                 match_start: None,
                                 snippet_deduplication_key: None,
                                 insert_text_mode: None,
                                 confirm: Some(Arc::new({
                                     let source = source.clone();
                                     move |intent, _window, cx| {
-                                        // Don't confirm errored commands
-                                        if is_error {
-                                            return false;
-                                        }
                                         if !is_missing_argument {
                                             cx.defer({
                                                 let source = source.clone();
