+++ crates/feature_flags/src/flags.rs
@@ -24,12 +24,6 @@ impl FeatureFlag for AcpBetaFeatureFlag {
     const NAME: &'static str = "acp-beta";
 }
 
-pub struct UserSlashCommandsFeatureFlag;
-
-impl FeatureFlag for UserSlashCommandsFeatureFlag {
-    const NAME: &'static str = "slash-commands";
-}
-
 pub struct ToolPermissionsFeatureFlag;
 
 impl FeatureFlag for ToolPermissionsFeatureFlag {
